+++ app/src/main/java/com/codesoom/assignment/dto/user/UserModificationData.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.dto;
+package com.codesoom.assignment.dto.user;
 
 import com.github.dozermapper.core.Mapping;
 import lombok.AllArgsConstructor;
@@ -13,12 +13,7 @@ import javax.validation.constraints.Size;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserRegistrationData {
-    @NotBlank
-    @Size(min = 3)
-    @Mapping("email")
-    private String email;
-
+public class UserModificationData {
     @NotBlank
     @Mapping("name")
     private String name;
